fix: inline createdAt anchor in copy message link
Address PR #1393 review: revert the messageLink.ts/ChatListUtils refactor and
keep only the minimal inline change in ChatBubble.onCopyLink. Copied message
links now append &createdAt=<epoch ms> so the receiver (ConversationPanelV2)
can anchor the scroll to the exact message instead of falling back to the
bottom of the channel for older, uncached messages.
